fix popup
