+++ pages/DishPopUpTest.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import DishPopup from '../components/DishPopUp/DishPopUp';
+
+const DishPopUpTest = () => {
+  return (
+    <div>
+      <DishPopup />
+    </div>
+  );
+};
+export default DishPopUpTest;
